phase_4/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic>=2.5.0
python-multipart==0.0.6
better-exceptions==0.3.3
openai>=1.3.7
python-dotenv==1.0.0
pydantic-settings==2.1.0
cryptography>=41.0.0
bcrypt>=4.0.0
alembic>=1.13.0
sqlalchemy>=2.0.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-slugify>=8.0.0
phonenumbers>=8.13.0
email-validator>=2.0.0
httpx>=0.25.0
requests>=2.31.0
psycopg2-binary>=2.9.0